feat(Battery): add interface for components with battery
