+++ include/openhaptics.h
@@ -4,7 +4,6 @@
 
 #include <abstract_component.hpp>
 #include <abstract_connection.hpp>
-#include <components_registry.hpp>
 #include <output.hpp>
 
 #if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
@@ -13,7 +12,11 @@
 
 #include <vector>
 
-class OpenHaptics final : public OH::IComponentRegistry<OH::AbstractComponent> {
+class OpenHaptics final : public OH::IComponentRegistry<OH::AbstractComponent>
+#if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
+, public OH::IBatteryConnected
+#endif
+{
  private:
   std::vector<OH::AbstractComponent*> components{};
   OH::AbstractConnection* connection;
