+++ lib/core/logging.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#if defined(ARDUINO_ARCH_ESP32)
+
 #include "esp_log.h"
 
 #define OH_LOGE(tag, format, ...) ESP_LOGE(tag, format, ##__VA_ARGS__)
@@ -7,3 +9,5 @@
 #define OH_LOGI(tag, format, ...) ESP_LOGI(tag, format, ##__VA_ARGS__)
 #define OH_LOGD(tag, format, ...) ESP_LOGD(tag, format, ##__VA_ARGS__)
 #define OH_LOGV(tag, format, ...) ESP_LOGV(tag, format, ##__VA_ARGS__)
+
+#endif
